Move new exception code into module, improve backcompat somewhat
